+++ duality/montecarlo/montecarlo.py
@@ -1,36 +1,80 @@
-#imports essential functions from the duality module.
-from duality.hub.hub import random_value
-from duality.misc.global_functions import *
-
-#shows detailed overview of available functions.
-def help():  
-    print('*** NOTE: duality.montecarlo works only for sequential data with reasonable standard deviation, otherwise simulated data would expand to infinity.\nIf exponential increase in data is detected, error is raised automatically. ***\n')
-    print('duality.montecarlo CALLABLE FUNCTIONS:\n')
-    print('.help() - information about the available functions in the duality.montecarlo module.\n * takes no additional arguments.\n')
-    print('.Configuration() - main class that defines the data, desired time sequence and number of simulations.\n * takes 4 additional arguments.\n   list_of_values - pandas dataframe of values.\n   time_seq - desired time sequence.\n   num_sims - desired number of simulation iterations.\n   log_summary (default: log_summary = False) - event log of executed functions.\n * Requirements:\n   pandas Python module\n   pd.DataFrame() function.')
-
 #object that contains the simulation data.
-class Configuration:
+class MonteCarlo:
 
-    #shows detailed overview of available functions. Performing this action will not be tracked in class logs.
-    def help():  
-        print('duality.montecarlo.Configuration CALLABLE FUNCTIONS:\n')
-        print('.help() - information about the available functions in the class.\n * takes no additional arguments.\n')
-        print('.execute() - executes a Monte Carlo simulation on a defined data set.\n * takes 1 optional argument (default: ref_value_index = 0)\n   ref_value_index - index of a reference value as a simulation starting point.\n * Requirements:\n   duality.montecarlo.Configuration().\n * Limitations:\n If exponential increase in data values is detected, automatically raises error.\n')
-        print('.graph() - plots the Monte Carlo simulation on a graph.\n * takes 4 optional customization arguments. (default: graph_title = \'Monte Carlo simulation\', x_title = \'X axis\', y_title = \'Y axis\', plot_size = (25,10).)\n   graph_title - title of the graph\n   x_title - title of the X axis.\n   y_title - title on the Y axis.\n   plot_size - desired size of the graph. eg. - (x_lenght_num, y_lenght_num). - NOTE: values must be inside the parentheses and divided by a comma.\n * Requirements:\n   duality.montecarlo.Configuration.execute().\n')
-        print('.get_risk() - calculates the risk of value decrease over time.\n * takes 1 optional argument (default: risk_sims = 5000).\n * Requirements:\n   duality.montecarlo.Configuration.execute().\n')
-        print('.get_stats() - shows the statistics of the Monte Carlo simulation.\n * takes no additional arguments.\n * Requirements:\n   duality.montecarlo.Configuration.execute().\n')
-        print('.get_logs() - shows the event log of executed functions.\n * takes no additional arguments.\n * Requirements:\n   duality.montecarlo.Configuration.execute().\n   log_summary = True.\n')
-        print('.get_change() - shows the percentage of Monte Carlo simulation value change for every iteration.\n * takes no additional arguments.\n * Requirements:\n   duality.montecarlo.Configuration.execute().\n')
-        print('.hist() - plots the histogram of Monte Carlo simulation.\n * takes 5 optional customization arguments. (default: graph_title = \'Monte Carlo simulation\', x_title = \'X axis\', y_title = \'Y axis\', plot_size = (25,10), method = \'b\'.)\nIf method = \'e\' is chosen, no customization arguments apply.\n   graph_title - title of the graph\n   x_title - title of the X axis.\n   y_title - title on the Y axis.\n   plot_size - desired size of the graph. eg. - (x_lenght_num, y_lenght_num). - NOTE: values must be inside the parentheses and divided by a comma.\n   method - default method is Basic histogram and it\'s performed by automation. In order to plot Empirical rule histogram add method = \'e\' as the last argument. - NOTE: method of a histogram must be placed within quotation marks.\n * Requirements:\n   duality.montecarlo.Configuration.execute().\n   duality.montecarlo.Configuration.get_stats().')
+    '''
+
+    (OBJECT INFO)
+    --- 
+    duality.MonteCarlo - main class that defines the data, desired time sequence and number of simulations.
+        * takes 5 additional arguments.
+            list_of_values - pandas dataframe of values.
+            time_seq - desired time sequence.
+            num_sims - desired number of simulation iterations.
+            log_summary (default: log_summary = False) - event log of executed functions.
+        * Requirements:
+            pandas Python module.
+            pd.DataFrame() defined data set.
+        * automatically executes the .execute() function.
+
+    duality.MonteCarlo CALLABLE FUNCTIONS:
+
+        .execute() - executes a Monte Carlo simulation on a defined data set.
+            *is automatically executed with the Object setup.
+
+        .graph() - plots the Monte Carlo simulation on a graph.
+            * takes 4 optional customization arguments. (default: graph_title = 'Monte Carlo simulation', x_title = 'X axis', y_title = 'Y axis', plot_size = (25,10)).
+                graph_title - title of the graph.
+                x_title - title of the X axis.
+                y_title - title on the Y axis.
+                plot_size - desired size of the graph. eg. - (x_lenght_num, y_lenght_num). - NOTE: values must be inside the parentheses and divided by a comma.
+
+        .get_risk() - calculates the risk of value decrease over time.
+            * takes 1 optional argument (default: risk_sims = 5000).
+
+        .get_stats() - shows the statistics of the Monte Carlo simulation.
+            * takes no additional arguments.
+
+        .get_logs() - shows the event log of executed functions.
+            * takes no additional arguments.
+            * Requirements:
+                log_summary = True.
+        
+        .get_change() - shows the percentage of Monte Carlo simulation value change for every iteration.
+            * takes no additional arguments.
+
+        .hist() - plots the histogram of Monte Carlo simulation.
+            * takes 5 optional customization arguments. (default: graph_title = 'Monte Carlo simulation', x_title = 'X axis', y_title = 'Y axis', plot_size = (25,10), method = 'b').
+            If method = 'e' is chosen, no customization arguments apply.
+                graph_title - title of the graph
+                x_title - title of the X axis.
+                y_title - title on the Y axis.
+                plot_size - desired size of the graph. eg. - (x_lenght_num, y_lenght_num). - NOTE: values must be inside the parentheses and divided by a comma.
+                method - default method is Basic histogram and it's performed by automation. In order to plot Empirical rule histogram add method = 'e' as the last argument. - NOTE: method of a histogram must be placed within quotation marks.
+            * automatically executes the .get_stats(filtered = True) function in order to get standard deviation for the Empirical rule plotting.
+
+    '''
+    
+    #metadata of the used package.
+    from duality.misc._meta import (
+        __author__,
+        __copyright__,
+        __credits__,
+        __license__,
+        __version__,
+        __documentation__,
+        __contact__,
+        __donate__
+    )
 
     #initial value configuration.
-    def __init__(self, list_of_values, time_seq, num_sims, log_summary = False): 
+    def __init__(self, list_of_values, time_seq, num_sims, ref_value_index = 0, log_summary = False): 
         self.list_of_values = list_of_values
         self.time_seq = time_seq
         self.num_sims = num_sims
         self.log_summary = log_summary
+        self.ref_value_index = ref_value_index
         print(f'Monte Carlo has been set up for {self.num_sims} simulations in a period of {self.time_seq} time measurement units.')
+        self.execute()
 
     #creates an event log that tracks the function execution time and duration.
     def classLog(func_name):
@@ -42,7 +86,7 @@ class Configuration:
                         start = time.time()
                         results = func(self, *args, **kwargs)
                         with open('duality Logs.txt', 'a') as f:
-                            f.write('Perfomed a function ' + func_name + ' at: ' + str(datetime.datetime.now()) + '.' + ' Time spent performing the action: ' + str(time.time() - start) + ' seconds.' + '\n')
+                            f.write('Performed a function ' + func_name + ' at: ' + str(datetime.datetime.now()) + '.' + ' Time spent performing the action: ' + str(time.time() - start) + ' seconds.' + '\n')
                             return results
                     else:
                             results = func(self, *args, **kwargs)
@@ -50,19 +94,20 @@ class Configuration:
                 return logsaver
         return log
         
-    #class information.
     @classLog('__str__()')
+    #class information.
     def __str__(self):
         return f'Monte Carlo defining object that stores the configuration data for creating {self.num_sims} simulations in a period of {self.time_seq} time measurement units.'
 
-    #class information.
     @classLog('__repr__()')
+    #class information.
     def __repr__(self):
         return f'Monte Carlo defining object that stores the configuration data for creating {self.num_sims} simulations in a period of {self.time_seq} time measurement units.'
 
+    @classLog('execute() - built in function.')
     #executes a Monte Carlo simulation on a defined data set.
-    @classLog('execute()')
-    def execute(self, ref_value_index = 0):
+    def execute(self):
+            from duality.hub.hub import random_value
             print('Monte Carlo simulation has been executed')
             print('NOTE: Use data with reasonable standard deviation in order to prevent exponential growth of the function that cannot be plotted properly, recognize such abnormal values by a + sign anywhere in the data executed below.\nThe model that will be able to handle big standard deviations is currently being worked on, thank you for your patience.\n')
             import pandas as pd
@@ -70,8 +115,7 @@ class Configuration:
             from warnings import simplefilter
             simplefilter(action="ignore", category=pd.errors.PerformanceWarning)
             #End of pandas warning removal block.
-            self.ref_value_index = ref_value_index
-            today_value = self.list_of_values.iloc[ref_value_index]
+            today_value = self.list_of_values.iloc[self.ref_value_index]
             data = pd.DataFrame()
             loading = 0
 
@@ -85,8 +129,7 @@ class Configuration:
                 if simulated_index > (index_array[-1] * 2):
                     raise Exception('Variation between data is too big, due to detection of exponentional increase of values or non-sequential data Monte Carlo simulation cannot be executed properly.')
                         
-                for num_day in range(self.time_seq):
-                    
+                for num_day in range(self.time_seq):   
                     rand_change = random_value(self.list_of_values.pct_change().mean(), self.list_of_values.pct_change().std())
                     if count == self.time_seq:
                         break
@@ -107,13 +150,13 @@ class Configuration:
             self.results = data
             return data
 
-    #shows the percentage of Monte Carlo simulation value change for every iteration.
     @classLog('get_change()')
+    #shows the percentage of Monte Carlo simulation value change for every iteration.
     def get_change(self):
         return self.results.pct_change()
    
-    #calculates the risk of negative values occuring.
     @classLog('get_risk()')
+    #calculates the risk of negative values occuring.
     def get_risk(self, risk_sims = 5000):
         import random
         import pandas as pd
@@ -143,8 +186,8 @@ class Configuration:
 
         print('Risk for this option is', round(NRisk,2), '%.')
      
-    #plots the Monte Carlo simulation on a graph.   
     @classLog('graph()')
+    #plots the Monte Carlo simulation on a graph.   
     def graph(self, graph_title = 'Monte Carlo simulation', x_title = 'X axis', y_title = 'Y axis', plot_size = (25,10)):
         print('MonteCarlo() plotting initialized.')
         import matplotlib.pyplot as plt
@@ -158,30 +201,33 @@ class Configuration:
         plt.show()
         print('MonteCarlo() plotting finished.')
    
-    #shows the statistics of the Monte Carlo simulation.
     @classLog('get_stats()')
-    def get_stats(self): 
+    #shows the statistics of the Monte Carlo simulation.
+    def get_stats(self, filtered = False): 
         import numpy as np
-        print('Number of simulations: ', self.time_seq)
-        print('Number of iterations: ', self.num_sims)
         mean_value = np.mean(self.results.loc[self.time_seq])
         mean_value = round((mean_value),2)
-        print('Mean value: ', mean_value)
         standard_deviation = round(np.std(self.results.loc[self.time_seq]),2)
         standard_deviation = round((standard_deviation),2)
-        print('Standard deviation: ', standard_deviation)
         maximum_value = np.max(self.results.loc[self.time_seq])
         maximum_value = round((maximum_value),2)
-        print('Maximum: ', maximum_value)
         minimum_value = np.min(self.results.loc[self.time_seq])
         minimum_value = round((minimum_value),2)
-        print('Minimum: ', minimum_value)
         self.standard_deviation = standard_deviation
         self.mean_value = mean_value
+        if filtered == False:
+            print('Number of simulations: ', self.time_seq)
+            print('Number of iterations: ', self.num_sims)
+            print('Mean value: ', mean_value)
+            print('Standard deviation: ', standard_deviation)
+            print('Maximum: ', maximum_value)
+            print('Minimum: ', minimum_value)
 
-    #plots the histogram of Monte Carlo simulation.
+
     @classLog('hist()')
+    #plots the histogram of Monte Carlo simulation.
     def hist(self, graph_title = 'Histogram of value frequencies', x_title = 'X axis', y_title = 'Y axis', plot_size = (25,10), **method):
+        self.get_stats(filtered = True)
         std_plus = self.mean_value + self.standard_deviation
         std_minus = self.mean_value - self.standard_deviation
         std_plus2 = self.mean_value + (self.standard_deviation * 2)
@@ -202,6 +248,7 @@ class Configuration:
             plt.suptitle(graph_title, fontsize = 25, weight = 'bold')
             
         if method.get("method") == "e":
+
             print('CHOSEN METHOD: Empirical rule.')
             plt.suptitle('Value division based on the Empirical rule', fontsize = 25, weight = 'bold')
             plt.axvline(x = std_plus, color = 'g', linestyle = 'dashed')
@@ -218,6 +265,7 @@ class Configuration:
         print('Histogram plotting finished.')
 
     @classLog('get_logs()')
+    #returns the saved logs of executed functions.
     def get_logs(self):
         f = open('duality Logs.txt', 'r')
         print(f.read())